+++ internal/pkg/process/analyze.go
@@ -19,12 +19,10 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"runtime"
 
 	"github.com/hashicorp/go-version"
 
 	"go.opentelemetry.io/auto/internal/pkg/log"
-	"go.opentelemetry.io/auto/internal/pkg/process/ptrace"
 )
 
 // TargetDetails are the details about a target function.
@@ -34,12 +32,6 @@ type TargetDetails struct {
 	GoVersion         *version.Version
 	Libraries         map[string]string
 	AllocationDetails *AllocationDetails
-}
-
-// AllocationDetails are the details about allocated memory.
-type AllocationDetails struct {
-	StartAddr uint64
-	EndAddr   uint64
 }
 
 // Func represents a function target.
@@ -78,44 +70,6 @@ func (t *TargetDetails) GetFunctionReturns(name string) ([]uint64, error) {
 	return nil, fmt.Errorf("could not find returns for function %s", name)
 }
 
-func (a *Analyzer) remoteMmap(pid int, mapSize uint64) (uint64, error) {
-	runtime.LockOSThread()
- 	defer runtime.UnlockOSThread()
-	program, err := ptrace.NewTracedProgram(pid, log.Logger)
-	if err != nil {
-		log.Logger.Error(err, "Failed to attach ptrace", "pid", pid)
-		return 0, err
-	}
-
-	defer func() {
-		log.Logger.V(0).Info("Detaching from process", "pid", pid)
-		err := program.Detach()
-		if err != nil {
-			log.Logger.Error(err, "Failed to detach ptrace", "pid", pid)
-		}
-	}()
-	fd := -1
-	addr, err := program.Mmap(mapSize, uint64(fd))
-	if err != nil {
-		log.Logger.Error(err, "Failed to mmap", "pid", pid)
-		return 0, err
-	}
-
-	err = program.Madvise(addr, mapSize)
-	if err != nil {
-		log.Logger.Error(err, "Failed to madvise", "pid", pid)
-		return 0, err
-	}
-
-	err = program.Mlock(addr, mapSize)
-	if err != nil {
-		log.Logger.Error(err, "Failed to mlock", "pid", pid)
-		return 0, err
-	}
-
-	return addr, nil
-}
-
 // Analyze returns the target details for an actively running process.
 func (a *Analyzer) Analyze(pid int, relevantFuncs map[string]interface{}) (*TargetDetails, error) {
 	result := &TargetDetails{
@@ -139,20 +93,6 @@ func (a *Analyzer) Analyze(pid int, relevantFuncs map[string]interface{}) (*Targ
 	}
 	result.GoVersion = goVersion
 	result.Libraries = modules
-
-	mapSize := uint64(os.Getpagesize() * runtime.NumCPU() * 8)
-	addr, err := a.remoteMmap(pid, mapSize)
-	if err != nil {
-		log.Logger.Error(err, "Failed to mmap")
-		return nil, err
-	}
-	log.Logger.V(0).Info("mmaped remote memory", "start_addr", fmt.Sprintf("%X", addr),
-		"end_addr", fmt.Sprintf("%X", addr+mapSize))
-
-	result.AllocationDetails = &AllocationDetails{
-		StartAddr: addr,
-		EndAddr:   addr + mapSize,
-	}
 
 	if err != nil {
 		return nil, err
